+++ standalone/src/main/java/software/xdev/mockserver/netty/HttpRequestHandler.java
@@ -25,8 +25,6 @@ import software.xdev.mockserver.dashboard.DashboardHandler;
 import software.xdev.mockserver.lifecycle.LifeCycle;
 import software.xdev.mockserver.log.model.LogEntry;
 import software.xdev.mockserver.logging.MockServerLogger;
-import software.xdev.mockserver.metrics.Metrics;
-import software.xdev.mockserver.metrics.MetricsHandler;
 import software.xdev.mockserver.mock.HttpState;
 import software.xdev.mockserver.mock.action.http.HttpActionHandler;
 import software.xdev.mockserver.model.HttpRequest;
@@ -53,7 +51,6 @@ import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import static software.xdev.mockserver.exception.ExceptionHandling.closeOnFlush;
 import static software.xdev.mockserver.exception.ExceptionHandling.connectionClosedException;
 import static software.xdev.mockserver.log.model.LogEntry.LogMessageType.AUTHENTICATION_FAILED;
-import static software.xdev.mockserver.metrics.Metrics.Name.REQUESTS_RECEIVED_COUNT;
 import static software.xdev.mockserver.mock.HttpState.PATH_PREFIX;
 import static software.xdev.mockserver.model.HttpResponse.response;
 import static software.xdev.mockserver.model.PortBinding.portBinding;
@@ -70,24 +67,20 @@ public class HttpRequestHandler extends SimpleChannelInboundHandler<HttpRequest>
     private final Configuration configuration;
     private LifeCycle server;
     private HttpState httpState;
-    private Metrics metrics;
     private MockServerLogger mockServerLogger;
     private PortBindingSerializer portBindingSerializer;
     private HttpActionHandler httpActionHandler;
     private DashboardHandler dashboardHandler;
-    private MetricsHandler metricsHandler;
 
     public HttpRequestHandler(Configuration configuration, LifeCycle server, HttpState httpState, HttpActionHandler httpActionHandler) {
         super(false);
         this.configuration = configuration;
         this.server = server;
         this.httpState = httpState;
-        this.metrics = new Metrics(configuration);
         this.mockServerLogger = httpState.getMockServerLogger();
         this.portBindingSerializer = new PortBindingSerializer(mockServerLogger);
         this.httpActionHandler = httpActionHandler;
         this.dashboardHandler = new DashboardHandler();
-        this.metricsHandler = new MetricsHandler(configuration);
     }
 
     private static boolean isProxyingRequest(ChannelHandlerContext ctx) {
@@ -114,10 +107,6 @@ public class HttpRequestHandler extends SimpleChannelInboundHandler<HttpRequest>
 
     @Override
     protected void channelRead0(final ChannelHandlerContext ctx, final HttpRequest request) {
-
-        if (configuration.metricsEnabled()) {
-            metrics.increment(REQUESTS_RECEIVED_COUNT);
-        }
 
         ResponseWriter responseWriter = new NettyResponseWriter(configuration, mockServerLogger, ctx, httpState.getScheduler());
         try {
@@ -154,10 +143,6 @@ public class HttpRequestHandler extends SimpleChannelInboundHandler<HttpRequest>
                 } else if (request.getMethod().getValue().equals("GET") && request.getPath().getValue().startsWith(PATH_PREFIX + "/dashboard")) {
 
                     dashboardHandler.renderDashboard(ctx, request);
-
-                } else if (configuration.metricsEnabled() && request.getMethod().getValue().equals("GET") && request.getPath().getValue().matches(PATH_PREFIX + "/metrics")) {
-
-                    metricsHandler.renderMetrics(ctx, request);
 
                 } else if (request.getMethod().getValue().equals("CONNECT")) {